Set time to 0 explicitly
Creating tasks would fail 1 in 10 times. The reasons for this is that
now would sometimes be one-second ahead. Therefore the task endpoint
would return a 400 for bad request. As the validator just needs to
check that now is either todays date or in the future I have removed
the time from the Date Time